requirements.txt
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.54
requests>=2.32.0
scikit-learn>=1.5.0
joblib>=1.4.0
matplotlib>=3.9.0
python-dotenv>=1.0.1
streamlit>=1.42.0
plotly>=5.24.0
